Minor FFT integration changes

+++ include/nil/algebra/fields/detail/element/fp.hpp
@@ -40,7 +40,7 @@ namespace nil {
 
                     element_fp(modulus_type data) : data(data, modulus) {};
 
-                    element_fp(size_t data) : data(data, modulus) {};
+                    element_fp(int data) : data(data, modulus) {};
 
                     element_fp(const element_fp &B) {
                         data = B.data;
